cli-package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archon-cli"
version = "0.1.0"
description = "ARCHON CLI - AI Architecture Review from your terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "ARCHON", email = "hello@archon.dev" }]
keywords = ["architecture", "review", "AI", "security", "cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
archon = "archon_cli.main:cli"

[project.urls]
Homepage = "https://archon.dev"
Repository = "https://github.com/VenkataAnilKumar/GenAI-Platform"
Documentation = "https://docs.archon.dev"
